+++ go.mod
@@ -4,7 +4,6 @@ go 1.18
 
 require (
 	github.com/gin-gonic/gin v1.7.7
-	github.com/rs/xid v1.3.0
 	go.mongodb.org/mongo-driver v1.8.4
 )
