+++ main.go
@@ -20,32 +20,16 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
+	handler "github.com/bunyawats/recipes-api/handlers"
+	"github.com/bunyawats/recipes-api/models"
 	"github.com/gin-gonic/gin"
-	"go.mongodb.org/mongo-driver/bson"
-	"go.mongodb.org/mongo-driver/bson/primitive"
 	"go.mongodb.org/mongo-driver/mongo"
 	"go.mongodb.org/mongo-driver/mongo/options"
 	"log"
-	"net/http"
 	"os"
-	"strings"
-	"time"
 )
 
-type Recipe struct {
-	ID           primitive.ObjectID `json:"id" bson:"_id"`
-	Name         string             `json:"name" bson:"name"`
-	Tags         []string           `json:"tags" bson:"tags"`
-	Ingredients  []string           `json:"ingredients" bson:"ingredients"`
-	Instructions []string           `json:"instructions" bson:"instructions"`
-	PublishedAt  time.Time          `json:"publishedAt" b`
-}
-
-var recipes []Recipe
-var ctx context.Context
-var err error
-var client *mongo.Client
-var collection *mongo.Collection
+var recipesHandler *handler.RecipesHandler
 
 func InitLoad() {
 
@@ -57,12 +41,12 @@ func InitLoad() {
 	fmt.Println("databaseName", databaseName)
 	fmt.Println("collectionName", collectionName)
 
-	recipes = make([]Recipe, 0)
+	recipes := make([]models.Recipe, 0)
 	file, _ := os.ReadFile("recipes.json")
 	_ = json.Unmarshal([]byte(file), &recipes)
 
-	ctx = context.Background()
-	client, err = mongo.Connect(
+	ctx := context.Background()
+	client, err := mongo.Connect(
 		ctx,
 		options.Client().ApplyURI(databaseUri),
 	)
@@ -90,237 +74,29 @@ func init() {
 	fmt.Println("databaseName", databaseName)
 	fmt.Println("collectionName", collectionName)
 
-	ctx = context.Background()
-	client, err = mongo.Connect(
+	ctx := context.Background()
+	client, err := mongo.Connect(
 		ctx,
 		options.Client().ApplyURI(databaseUri),
 	)
 	if err != nil {
 		log.Fatal("Connect to MongoDB failed:", err.Error())
 	}
-	collection = client.Database(databaseName).Collection(collectionName)
+	collection := client.Database(databaseName).Collection(collectionName)
 	log.Println("Connected to MongoDB")
-}
 
-// swagger:operation GET /recipes recipes listRecipes
-// Returns list of recipes
-// ---
-// produces:
-// - application/json
-// responses:
-//     '200':
-//         description: Successful operation
-func ListRecipesHandler(c *gin.Context) {
-
-	cur, err := collection.Find(ctx, bson.M{})
-	if err != nil {
-		log.Println("error: ", err.Error())
-		c.JSON(http.StatusInternalServerError,
-			gin.H{
-				"error": err.Error(),
-			},
-		)
-		return
-	}
-	defer cur.Close(ctx)
-
-	recipes := make([]Recipe, 0)
-	for cur.Next(ctx) {
-		var recipe Recipe
-		cur.Decode(&recipe)
-		recipes = append(recipes, recipe)
-	}
-	c.JSON(http.StatusOK, recipes)
-}
-
-// swagger:operation POST /recipes recipes newRecipe
-// Update an existing recipe
-// ---
-// parameters:
-// - name: id
-//   in: path
-//   description: ID of the recipe
-//   required: true
-//   type: string
-// produces:
-// - application/json
-// responses:
-//     '200':
-//         description: Successful operation
-//     '400':
-//         description: Invalid input
-//     '404':
-//         description: Invalid recipe ID
-func NewRecipeHandler(c *gin.Context) {
-	// validate request
-	var recipe Recipe
-	if err := c.ShouldBindJSON(&recipe); err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{
-			"error": err.Error(),
-		})
-		return
-	}
-	// insert to database
-	recipe.ID = primitive.NewObjectID()
-	recipe.PublishedAt = time.Now()
-	_, err := collection.InsertOne(ctx, recipe)
-	// response the result
-	if err != nil {
-		fmt.Println(err)
-		c.JSON(http.StatusInternalServerError, gin.H{
-			"error": "Error while inserting a new recipe",
-		})
-		return
-	}
-	c.JSON(http.StatusOK, recipe)
-}
-
-// swagger:operation PUT /recipes/{id} recipes updateRecipe
-// Update an existing recipe
-// ---
-// parameters:
-// - name: id
-//   in: path
-//   description: ID of the recipe
-//   required: true
-//   type: string
-// produces:
-// - application/json
-// responses:
-//     '200':
-//         description: Successful operation
-//     '400':
-//         description: Invalid input
-//     '404':
-//         description: Invalid recipe ID
-func UpdateRecipeHandler(c *gin.Context) {
-	// validate request
-	id := c.Param("id")
-	var recipe Recipe
-	if err := c.ShouldBindJSON(&recipe); err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{
-			"error": err.Error(),
-		})
-		return
-	}
-	// update to database
-	objectId, _ := primitive.ObjectIDFromHex(id)
-	_, err = collection.UpdateOne(
-		ctx,
-		bson.M{
-			"_id": objectId,
-		},
-		bson.D{
-			{
-				"$set", bson.D{
-					{"name", recipe.Name},
-					{"instructions", recipe.Instructions},
-					{"ingredients", recipe.Ingredients},
-					{"tags", recipe.Tags},
-				},
-			},
-		},
-	)
-	// response the result
-	if err != nil {
-		fmt.Println(err)
-		c.JSON(http.StatusInternalServerError, gin.H{
-			"error": err.Error(),
-		})
-		return
-	}
-	c.JSON(http.StatusOK, gin.H{
-		"message": "Recipe has been updated",
-	})
-}
-
-// swagger:operation DELETE /recipes/{id} recipes deleteRecipe
-// Update an existing recipe
-// ---
-// parameters:
-// - name: id
-//   in: path
-//   description: ID of the recipe
-//   required: true
-//   type: string
-// produces:
-// - application/json
-// responses:
-//     '200':
-//         description: Successful operation
-//     '400':
-//         description: Invalid input
-//     '404':
-//         description: Invalid recipe ID
-func DeleteRecipesHandler(c *gin.Context) {
-	// validate request
-	id := c.Param("id")
-	objectId, err := primitive.ObjectIDFromHex(id)
-	if err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{
-			"error": err.Error(),
-		})
-	}
-	// delete from database
-	_, err = collection.DeleteOne(ctx, bson.M{
-		"_id": objectId,
-	})
-	// response the result
-	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{
-			"error": err.Error(),
-		})
-		return
-	}
-	c.JSON(http.StatusOK, gin.H{
-		"message": "Recipe has been deleted",
-	})
-
-}
-
-// swagger:operation GET /recipes/search recipes searchRecipe
-// Update an existing recipe
-// ---
-// parameters:
-// - name: id
-//   in: path
-//   description: ID of the recipe
-//   required: true
-//   type: string
-// produces:
-// - application/json
-// responses:
-//     '200':
-//         description: Successful operation
-//     '400':
-//         description: Invalid input
-//     '404':
-//         description: Invalid recipe ID
-func SearchRecipesHandler(c *gin.Context) {
-	tag := c.Query("tag")
-	listOfRecipes := make([]Recipe, 0)
-	for i := 0; i < len(recipes); i++ {
-		found := false
-		for _, t := range recipes[i].Tags {
-
-			if strings.EqualFold(t, tag) {
-				found = true
-			}
-		}
-		if found {
-			listOfRecipes = append(listOfRecipes, recipes[i])
-		}
-	}
-	c.JSON(http.StatusOK, listOfRecipes)
-
+	recipesHandler = handler.NewRecipesHandler(ctx, collection)
 }
 
 func main() {
 	router := gin.Default()
-	router.POST("/recipes", NewRecipeHandler)
-	router.GET("/recipes", ListRecipesHandler)
-	router.PUT("/recipes/:id", UpdateRecipeHandler)
-	router.DELETE("/recipes/:id", DeleteRecipesHandler)
+	router.POST("/recipes", recipesHandler.NewRecipeHandler)
+	router.GET("/recipes", recipesHandler.ListRecipesHandler)
+	router.PUT("/recipes/:id", recipesHandler.UpdateRecipeHandler)
+	router.DELETE("/recipes/:id", recipesHandler.DeleteRecipesHandler)
 	//	router.GET("/recipes/search", SearchRecipesHandler)
-	router.Run()
+	err := router.Run()
+	if err != nil {
+		log.Fatal(err)
+	}
 }
